+++ src/shapes/RPoly.js
@@ -1,5 +1,6 @@
 /**
- * @classdesc 可以进行规则多边形填充或线框绘制的显示对象
+ * @classdesc 可以进行规则多边形填充或线框绘制的显示对象。该多边形拥有内外两个半径，
+ * 可以构成有趣的形状。外半径由对象的w属性决定，内半径则需要指定r属性
  * @class 
  * @extends soya2d.DisplayObjectContainer
  * @param {Object} data 所有父类参数,以及新增参数
@@ -7,8 +8,7 @@
  * @param {String} data.strokeStyle 线框样式
  * @param {String} data.lineWidth 线条宽度
  * @param {int} data.edgeCount 多边形的边数，不能小于3
- * @param {Number} data.r1 半径1
- * @param {Number} data.r2 半径2
+ * @param {Number} [data.r] 内半径。默认和外半径相同
  * @author {@link http://weibo.com/soya2d MrSoya}
  */
 soya2d.RPoly = function(data){
@@ -23,7 +23,7 @@ soya2d.ext(soya2d.RPoly.prototype,{
     onRender:function(g){
         g.beginPath();
         g.fillStyle(this.fillStyle);
-        g.regularPolygon(this.w/2,this.h/2,this.edgeCount,this.r1,this.r2);
+        g.regularPolygon(this.w/2,this.h/2,this.edgeCount,this.r||this.w/2,this.w/2);
         g.closePath();
         g.fill();
         if(this.lineWidth>0){