+++ src/shapes/RRect.js
@@ -15,11 +15,14 @@ soya2d.RRect = function(data){
 	soya2d.ext(this,data);
 
     this.fillStyle = data.fillStyle || 'transparent';
+
+    this.r = data.r || 0;
 };
 soya2d.inherits(soya2d.RRect,soya2d.DisplayObjectContainer);
 soya2d.ext(soya2d.RRect.prototype,{
     onRender:function(g){
         g.beginPath();
+        g.moveTo(0,0);
         g.fillStyle(this.fillStyle);
         g.roundRect(0,0,this.w,this.h,this.r);
         g.fill();